refactor: Move ts files folder related to upload
