+++ src/cake/entities/cake.schema.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
 import { Document, SchemaTypes } from 'mongoose';
-import { image } from '../../common/type/image.type';
+import { image } from '../../upload/type/image.type';
 import * as mongoosePaginate from 'mongoose-paginate-v2';
 import { Order } from '../../order/entities/order.schema';
 import { User } from '../../user/entities/user.schema';
