+++ src/upload/dto/image-response.dto.ts
